modularize schema
